chore: Replace old implementations with refactored code
- Replaced base_agent.py with refactored version using TaskStore
- Fixed imports and implementation details for working system
- Archived old base_agent.py for reference
- Kept MockOperationsService as primary service (works in mock mode)
- Added full CLI execution logic to refactored base agent

The system now uses:
- Centralized TaskStore instead of shared class variables
- Clean dependency injection in BaseAgentProvider
- Proper async/await patterns throughout
- ComputeProviderFactory for extensible routing

Note: Full operations.py refactoring postponed as current
MockOperationsService works well for the PoC use case.
