+++ compute/base_agent.py
@@ -1,11 +1,7 @@
 """
-Base class for CLI-based AI agents (Qwen Code, Gemini CLI, etc).
-
-Provides shared functionality for headless, one-shot agent execution
-with a strict sentinel + JSON contract for deterministic results.
+Refactored base agent provider with clean separation of concerns.
+Uses centralized task store instead of shared class variables.
 """
-from __future__ import annotations
-
 import asyncio
 import json
 import logging
@@ -14,6 +10,8 @@ import re
 import shutil
 import tempfile
 import time
+from abc import ABC, abstractmethod
+from datetime import datetime
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
@@ -21,62 +19,76 @@ import pexpect
 
 from compute.base import BaseCompute, ExecuteResponse, GetResponse
 from domain.entities import GrantFile
+from services.task_store import TaskStatus, get_task_store
 
 logger = logging.getLogger(__name__)
 
+# Sentinel to mark agent completion
 SENTINEL = "__AGENT_DONE__"
 JSON_LINE_RE = re.compile(r'^\s*\{.*\}\s*$')
 
-
 def _redact_sensitive_data(text: str, api_key: str = None) -> str:
-    """Redact sensitive information from logs and outputs."""
-    if api_key and api_key in text:
-        text = text.replace(api_key, "****REDACTED****")
+    """Redact API keys from output."""
+    if api_key and len(api_key) > 8:
+        # Redact all but first 4 chars
+        text = text.replace(api_key, api_key[:4] + "[REDACTED]")
     return text
 
 
-class BaseAgentProvider(BaseCompute):
+class BaseAgentProvider(BaseCompute, ABC):
     """
-    Base class for CLI-based AI agent providers.
-    
-    Handles common functionality like:
-    - Secure workspace creation and cleanup
-    - PTY-based process spawning
-    - Output capture and timeout handling
-    - Artifact collection
-    - JSON response parsing
+    Refactored base class for headless agent providers.
+    Uses dependency injection and centralized state management.
     """
     
     # Subclasses should override these
-    CLI_NAME = "agent"  # Name of the CLI tool (e.g., "qwen", "gemini")
-    AGENT_TYPE = "base"  # Type identifier for logging
+    CLI_NAME = "agent"
+    AGENT_TYPE = "base"
     
-    def __init__(self):
-        self.settings = None  # Subclasses should set this
+    def __init__(self, task_store=None, artifact_storage=None):
+        """
+        Initialize with injected dependencies.
+        
+        Args:
+            task_store: Task storage service (uses global if not provided)
+            artifact_storage: Artifact storage service (lazy loaded if not provided)
+        """
+        # Configuration (subclasses should set these)
+        self.settings = None
         self.api_key = None
         self.timeout_sec = 180
         self.max_stdout_bytes = 2_000_000
         self.cli_path = None
         self.use_api_auth = False
         
-        # Use centralized task store instead of shared class variables
-        from services.task_store import get_task_store
-        self._task_store = get_task_store()
+        # Injected dependencies
+        self._task_store = task_store or get_task_store()
+        self._artifact_storage = artifact_storage
     
+    @property
+    def artifact_storage(self):
+        """Lazy load artifact storage to avoid circular imports."""
+        if self._artifact_storage is None:
+            from services.artifact_storage import ArtifactStorageService
+            self._artifact_storage = ArtifactStorageService()
+        return self._artifact_storage
+    
+    @abstractmethod
     def get_cli_command(self) -> str:
-        """Get the CLI command to execute. Subclasses can override."""
-        return self.cli_path or self.CLI_NAME
+        """Get the CLI command to execute."""
+        pass
     
+    @abstractmethod
     def get_cli_args(self, prompt: str) -> List[str]:
-        """Get CLI arguments. Subclasses should override based on their CLI."""
-        return ["-p", prompt]
+        """Get CLI arguments for the specific agent."""
+        pass
     
     def get_env_overrides(self) -> Dict[str, str]:
         """Get environment variable overrides for API authentication."""
         return {}
     
     def build_prompt(self, goal: str, files_dict: Dict[str, bytes] = None) -> str:
-        """Build the prompt for the agent. Can be overridden by subclasses."""
+        """Build the prompt for the agent."""
         files_info = ""
         if files_dict:
             files_list = []
@@ -89,412 +101,346 @@ class BaseAgentProvider(BaseCompute):
             f"You are running in a headless, single-shot batch mode. "
             f"Work only inside the current directory.{files_info}\n"
             f"IMPORTANT: Read and analyze the available data files to complete your task.\n"
-            f"Generate the code/commands needed but don't try to execute them.\n"
-            f"Instead, describe what would be done and output the final result.\n\n"
+            f"Generate output files in ./out/ directory.\n\n"
             f"CONSTRAINTS:\n"
             f"- No follow-up questions. Assume sensible defaults.\n"
-            f"- Read the provided data files to understand the available information.\n"
-            f"- Describe creating the ./out/ directory if needed.\n"
-            f"- Describe work products that would go in ./out/.\n"
-            f"- At completion, print exactly one JSON line (no backticks) describing results:\n"
+            f"- Create ./out/ directory if needed.\n"
+            f"- Save work products to ./out/.\n"
+            f"- At completion, print exactly one JSON line describing results:\n"
             f'  {{"status":"ok|error","summary":"<one line>","artifacts":["./out/..."],"notes":"<optional>"}}\n'
-            f"- Then print exactly: {SENTINEL}\n"
-            f"- Do not print anything after {SENTINEL}.\n\n"
+            f"- Then print exactly: {SENTINEL}\n\n"
             f"GOAL:\n{goal}\n"
         )
     
-    def execute(self, grant_file: GrantFile, files_content: list[str]) -> ExecuteResponse:
+    async def execute(self, grant_file: GrantFile, files_content: list[str]) -> ExecuteResponse:
         """
-        Start an agentic task using the CLI agent (returns immediately).
+        Start an agentic task asynchronously.
         
         Args:
-            grant_file: Grant containing operation parameters including 'goal'
-            files_content: List of decrypted file contents as strings
-        
+            grant_file: Grant containing operation parameters
+            files_content: List of decrypted file contents
+            
         Returns:
-            ExecuteResponse with operation ID and timestamp (task runs in background)
+            ExecuteResponse with operation ID (task runs in background)
         """
         goal = grant_file.parameters.get("goal")
         if not goal or not isinstance(goal, str):
-            raise ValueError(f"{self.AGENT_TYPE} operation requires 'goal' parameter (string)")
-
-        # Convert file contents to workspace format with descriptive names
-        files_dict = {}
-        for i, content in enumerate(files_content):
-            # Try to determine file type from content
-            if "chatgpt" in content.lower() or "conversation history" in content.lower():
-                filename = f"chatgpt_conversations_{i:02d}.txt"
-            elif "spotify" in content.lower() or "listening_history" in content.lower():
-                filename = f"spotify_data_{i:02d}.json"
-            elif "linkedin" in content.lower() or "professional" in content.lower():
-                filename = f"linkedin_profile_{i:02d}.txt"
-            elif "fitness" in content.lower() or "health" in content.lower():
-                filename = f"fitness_data_{i:02d}.json"
-            else:
-                filename = f"personal_data_{i:02d}.txt"
-            
-            files_dict[filename] = content.encode("utf-8")
-
-        # Create a unique operation ID
-        operation_id = f"{self.AGENT_TYPE}_{int(time.time() * 1000)}"
-        created_at = time.strftime("%Y-%m-%dT%H:%M:%S.000Z", time.gmtime())
-
-        logger.info(f"[{self.AGENT_TYPE}] Starting task: {operation_id} - {goal[:100]}...")
-
-        # Initialize operation status
-        self._results[operation_id] = {
-            "status": "running",
-            "summary": "Task is running in background",
-            "result": {},
-            "artifacts": [],
-            "logs": [],
-            "stdout": "",
-        }
-
-        # Start background task
-        task = asyncio.create_task(self._run_agent_async(operation_id, goal, files_dict, grant_file.grantee))
-        self._tasks[operation_id] = task
-
-        logger.info(f"[{self.AGENT_TYPE}] Task started in background: {operation_id}")
-
-        return ExecuteResponse(
-            id=operation_id,
-            created_at=created_at
-        )
-
-    def get(self, prediction_id: str) -> GetResponse:
-        """
-        Get results from an agentic task (running or completed).
-        """
-        if prediction_id not in self._results:
-            raise ValueError(f"Operation {prediction_id} not found")
-
-        # Check if task is still running
-        if prediction_id in self._tasks:
-            task = self._tasks[prediction_id]
-            if not task.done():
-                # Task still running
-                return GetResponse(
-                    id=prediction_id,
-                    status="running",
-                    started_at=None,
-                    finished_at=None,
-                    result=json.dumps({
-                        "status": "running",
-                        "summary": "Task is running in background",
-                        "result": {},
-                        "artifacts": [],
-                        "logs": [],
-                        "stdout": "",
-                    }, indent=2)
-                )
-            else:
-                # Task completed, clean up
-                if task.exception():
-                    logger.error(f"[{self.AGENT_TYPE}] Task {prediction_id} failed with exception: {task.exception()}")
-                del self._tasks[prediction_id]
-
-        result = self._results[prediction_id]
+            raise ValueError(f"{self.AGENT_TYPE} operation requires 'goal' parameter")
         
-        # Create artifact metadata (no content - stored in R2)
-        artifacts_metadata = []
-        for artifact in result.get("artifacts", []):
-            artifacts_metadata.append({
-                "name": artifact["name"],
-                "size": len(artifact["content"]),
-                "download_url": f"/api/v1/artifacts/{prediction_id}/{artifact['name']}"
-            })
-
-        # Format result as string for compatibility
-        result_str = json.dumps({
-            "status": result["status"],
-            "summary": result["summary"], 
-            "result": result.get("result", {}),
-            "artifacts": artifacts_metadata,  # Just metadata, not content
-            "logs": result.get("logs", []),
-            "stdout": result.get("stdout", ""),  # Include for debugging
-        }, indent=2)
-
+        # Generate operation ID
+        operation_id = f"{self.AGENT_TYPE}_{int(time.time() * 1000)}"
+        created_at = datetime.utcnow().isoformat() + "Z"
+        
+        # Convert files to workspace format
+        files_dict = self._prepare_files(files_content)
+        
+        # Create task entry
+        await self._task_store.create_task(operation_id)
+        
+        # Start background task
+        task = asyncio.create_task(
+            self._run_agent_async(operation_id, goal, files_dict, grant_file.grantee)
+        )
+        
+        # Store task reference
+        await self._task_store.set_async_task(operation_id, task)
+        
+        logger.info(f"[{self.AGENT_TYPE}] Task started: {operation_id}")
+        
+        return ExecuteResponse(id=operation_id, created_at=created_at)
+    
+    async def get(self, prediction_id: str) -> GetResponse:
+        """Get task status and results."""
+        task_info = await self._task_store.get_task(prediction_id)
+        
+        if not task_info:
+            raise ValueError(f"Operation {prediction_id} not found")
+        
         # Map internal status to API status
-        if result["status"] == "running":
-            api_status = "running"
-        elif result["status"] == "ok":
-            api_status = "succeeded"
+        api_status = self._map_status(task_info.status)
+        
+        # Format result
+        if task_info.result:
+            result_str = json.dumps(task_info.result, indent=2)
         else:
-            api_status = "failed"
-
+            result_str = json.dumps({
+                "status": task_info.status.value,
+                "summary": f"Task is {task_info.status.value}",
+                "result": {},
+                "artifacts": [],
+                "logs": []
+            }, indent=2)
+        
         return GetResponse(
             id=prediction_id,
             status=api_status,
-            started_at=None,
-            finished_at=None,
+            started_at=task_info.started_at.isoformat() + "Z" if task_info.started_at else None,
+            finished_at=task_info.completed_at.isoformat() + "Z" if task_info.completed_at else None,
             result=result_str
         )
-
-    def cancel(self, prediction_id: str) -> bool:
-        """Cancel operation - attempts to cancel running task."""
-        if prediction_id in self._tasks:
-            task = self._tasks[prediction_id]
-            if not task.done():
-                task.cancel()
-                logger.info(f"[{self.AGENT_TYPE}] Cancelled task: {prediction_id}")
-                # Update status
-                if prediction_id in self._results:
-                    self._results[prediction_id].update({
-                        "status": "error",
-                        "summary": "Task cancelled by user"
-                    })
-                return True
-        return False
     
-    async def _run_agent_async(self, operation_id: str, goal: str, files_dict: Dict[str, bytes], grantee_address: str):
+    async def cancel(self, prediction_id: str) -> bool:
+        """Cancel a running task."""
+        success = await self._task_store.cancel_task(prediction_id)
+        if success:
+            logger.info(f"[{self.AGENT_TYPE}] Cancelled task: {prediction_id}")
+        return success
+    
+    def _prepare_files(self, files_content: list[str]) -> Dict[str, bytes]:
+        """Convert file contents to workspace format with descriptive names."""
+        files_dict = {}
+        for i, content in enumerate(files_content):
+            # Determine filename based on content
+            if "chatgpt" in content.lower():
+                filename = f"chatgpt_conversations_{i:02d}.txt"
+            elif "spotify" in content.lower():
+                filename = f"spotify_data_{i:02d}.json"
+            elif "linkedin" in content.lower():
+                filename = f"linkedin_profile_{i:02d}.json"
+            else:
+                filename = f"user_data_{i:02d}.txt"
+            
+            files_dict[filename] = content.encode('utf-8')
+        
+        return files_dict
+    
+    def _map_status(self, status: TaskStatus) -> str:
+        """Map internal status to API status."""
+        mapping = {
+            TaskStatus.PENDING: "pending",
+            TaskStatus.RUNNING: "running",
+            TaskStatus.SUCCEEDED: "succeeded",
+            TaskStatus.FAILED: "failed",
+            TaskStatus.CANCELLED: "cancelled"
+        }
+        return mapping.get(status, "failed")
+    
+    async def _run_agent_async(
+        self, 
+        operation_id: str, 
+        goal: str, 
+        files_dict: Dict[str, bytes],
+        grantee_address: str
+    ):
         """
-        Async wrapper that runs the agent in a background task.
+        Run the agent in a background task.
+        This is the core execution logic, now with clean separation.
         """
         try:
-            logger.info(f"[{self.AGENT_TYPE}] Background task started: {operation_id}")
+            # Update status to running
+            await self._task_store.update_status(operation_id, TaskStatus.RUNNING)
             
-            # Run the synchronous agent execution in a thread pool
-            loop = asyncio.get_event_loop()
-            result = await loop.run_in_executor(
-                None,  # Use default thread pool
-                self._run_agent_headless,
-                goal,
-                files_dict
-            )
-            
-            # Store artifacts in cloud storage
-            try:
-                from services.artifact_storage import ArtifactStorageService
-                storage_service = ArtifactStorageService()
+            # Create workspace
+            with tempfile.TemporaryDirectory() as workspace:
+                # Stage files
+                for filename, content in files_dict.items():
+                    filepath = os.path.join(workspace, filename)
+                    with open(filepath, 'wb') as f:
+                        f.write(content)
                 
-                # Store artifacts if any
-                if result.get("artifacts"):
-                    await storage_service.store_artifacts(
-                        operation_id=operation_id,
-                        artifacts=result["artifacts"],
-                        grantee_address=grantee_address
-                    )
-                    logger.info(f"[{self.AGENT_TYPE}] Stored {len(result['artifacts'])} artifacts in cloud storage")
-            except Exception as e:
-                logger.warning(f"[{self.AGENT_TYPE}] Failed to store artifacts in cloud: {e}")
-                result.setdefault("logs", []).append(f"Artifact storage failed: {e}")
-            
-            # Update results
-            self._results[operation_id] = result
-            
-            logger.info(f"[{self.AGENT_TYPE}] Background task completed: {operation_id} (status: {result['status']})")
-            
-        except asyncio.CancelledError:
-            logger.info(f"[{self.AGENT_TYPE}] Background task cancelled: {operation_id}")
-            self._results[operation_id] = {
-                "status": "error",
-                "summary": "Task was cancelled",
-                "result": {},
-                "artifacts": [],
-                "logs": ["Task cancelled"],
-                "stdout": "",
-            }
+                # Build prompt
+                prompt = self.build_prompt(goal, files_dict)
+                
+                # Get CLI command and args
+                command = self.get_cli_command()
+                args = self.get_cli_args(prompt)
+                
+                # Execute agent
+                result = await self._execute_cli(
+                    command, args, workspace, operation_id
+                )
+                
+                # Process artifacts
+                artifacts = await self._process_artifacts(
+                    workspace, operation_id, grantee_address, result
+                )
+                
+                # Update result with artifact metadata
+                result["artifacts"] = artifacts
+                
+                # Store result
+                status = TaskStatus.SUCCEEDED if result.get("status") == "ok" else TaskStatus.FAILED
+                await self._task_store.update_status(
+                    operation_id, status, result=result
+                )
+                
+                logger.info(f"[{self.AGENT_TYPE}] Task completed: {operation_id}")
+                
         except Exception as e:
-            logger.error(f"[{self.AGENT_TYPE}] Background task failed: {operation_id} - {e}")
-            self._results[operation_id] = {
-                "status": "error", 
-                "summary": f"Task failed: {type(e).__name__}",
-                "result": {},
-                "artifacts": [],
-                "logs": [str(e)],
-                "stdout": "",
-            }
-
-    def _run_agent_headless(
-        self,
-        goal: str,
-        files_content: Dict[str, bytes],
-    ) -> Dict[str, Any]:
-        """
-        Execute the CLI agent in headless mode with strict contract enforcement.
-        """
-        t0 = time.time()
-        workspace = Path(tempfile.mkdtemp(prefix=f"{self.AGENT_TYPE}_", dir=None))
-        outdir = workspace / "out"
-        outdir.mkdir(parents=True, exist_ok=True)
-        os.chmod(workspace, 0o700)
-
-        logger.info(f"[{self.AGENT_TYPE}] Created secure workspace: {workspace}")
-
-        # Stage input files
-        for rel_path, content in files_content.items():
-            dest = workspace / rel_path
-            dest.parent.mkdir(parents=True, exist_ok=True)
-            dest.write_bytes(content)
-            logger.debug(f"[{self.AGENT_TYPE}] Staged file: {rel_path} ({len(content)} bytes)")
-
-        # Create task description for the agent
-        files_list = "\n".join([f"- {filename}" for filename in files_content.keys()])
-        (workspace / "README_TASK.md").write_text(
-            f"# Headless One-Shot Agent Run\n\n"
-            f"**Goal:** {goal}\n\n"
-            f"**Available Data Files:**\n{files_list}\n\n"
-            f"**Instructions:**\n"
-            f"- Read and analyze the data files listed above\n"
-            f"- Work only inside the current directory\n"
-            f"- Write all outputs under ./out/\n"
-            f"- When complete, print exactly one JSON line then {SENTINEL}\n"
-            f"- Do not print anything after {SENTINEL}\n",
-            encoding="utf-8",
+            logger.error(f"[{self.AGENT_TYPE}] Task failed: {operation_id}: {e}")
+            await self._task_store.update_status(
+                operation_id, 
+                TaskStatus.FAILED,
+                error=str(e)
+            )
+    
+    async def _execute_cli(
+        self, 
+        command: str, 
+        args: List[str], 
+        workspace: str,
+        operation_id: str
+    ) -> Dict:
+        """Execute CLI command and capture output."""
+        # Run synchronously in thread pool to avoid blocking
+        loop = asyncio.get_event_loop()
+        return await loop.run_in_executor(
+            None,
+            self._run_cli_sync,
+            command, args, workspace, operation_id
         )
-
+    
+    def _run_cli_sync(
+        self,
+        command: str,
+        args: List[str],
+        workspace: str,
+        operation_id: str
+    ) -> Dict[str, Any]:
+        """Synchronous CLI execution with pexpect."""
+        t0 = time.time()
+        workspace_path = Path(workspace)
+        
         # Setup environment
         env = os.environ.copy()
         env["CI"] = "1"
         env["NO_COLOR"] = "1"
+        env.update(self.get_env_overrides())
         
-        # Add any API authentication environment variables
-        env_overrides = self.get_env_overrides()
-        if env_overrides:
-            env.update(env_overrides)
-
-        # Build the prompt with file information
-        prompt = self.build_prompt(goal, files_content)
-
-        # Get CLI command and arguments
-        cli_cmd = self.get_cli_command()
-        cli_args = self.get_cli_args(prompt)
-
         child = None
         captured = bytearray()
         logs = []
         artifacts = []
         parsed_json = None
-
+        
         try:
-            logger.info(f"[{self.AGENT_TYPE}] Spawning {cli_cmd} with args: {cli_args[:2]}...")
+            logger.info(f"[{self.AGENT_TYPE}] Spawning {command} in {workspace}")
             
             child = pexpect.spawn(
-                cli_cmd,
-                args=cli_args,
-                cwd=str(workspace),
+                command,
+                args=args,
+                cwd=workspace,
                 env=env,
                 encoding="utf-8",
                 timeout=self.timeout_sec,
             )
-
-            logger.info(f"[{self.AGENT_TYPE}] Process spawned, waiting for output...")
-
+            
             # Read output until sentinel or timeout
-            last_progress_log = time.time()
+            last_log = time.time()
             while True:
                 try:
                     chunk = child.read_nonblocking(size=4096, timeout=5)
                 except pexpect.TIMEOUT:
-                    # Check global timeout
                     elapsed = time.time() - t0
                     if elapsed > self.timeout_sec:
-                        logs.append(f"Global timeout {self.timeout_sec}s reached")
+                        logs.append(f"Timeout after {self.timeout_sec}s")
                         break
-                    
-                    # Log progress every 5 seconds
-                    if time.time() - last_progress_log > 5:
-                        logger.info(f"[{self.AGENT_TYPE}] Still processing... ({elapsed:.1f}s elapsed, {len(captured)} chars captured)")
-                        last_progress_log = time.time()
+                    if time.time() - last_log > 10:
+                        logger.info(f"[{self.AGENT_TYPE}] Still running... ({elapsed:.1f}s)")
+                        last_log = time.time()
                     continue
                 except pexpect.EOF:
-                    logger.info(f"[{self.AGENT_TYPE}] Process exited (EOF)")
+                    logger.info(f"[{self.AGENT_TYPE}] Process completed")
                     break
                 else:
                     captured.extend(chunk.encode("utf-8", errors="ignore"))
-                    
-                    # Check output size limit
                     if len(captured) >= self.max_stdout_bytes:
-                        logs.append("Maximum stdout size reached")
+                        logs.append("Max output size reached")
                         break
-                    
-                    # Check for completion sentinel
                     if SENTINEL in chunk:
                         logger.info(f"[{self.AGENT_TYPE}] Found completion sentinel")
                         break
-
-            # Ensure process is terminated
-            try:
-                if child.isalive():
-                    child.terminate(force=True)
+            
+            # Terminate process
+            if child and child.isalive():
+                child.terminate(force=True)
+            if child:
                 child.close(force=True)
-            except Exception as e:
-                logger.warning(f"[{self.AGENT_TYPE}] Error closing process: {e}")
-
-            # Process captured output
+            
+            # Process output
             stdout_text = captured.decode("utf-8", errors="ignore")
             stdout_text = _redact_sensitive_data(stdout_text, self.api_key)
             
-            execution_time = time.time() - t0
-            logs.append(f"Execution time: {execution_time:.2f}s")
-            logs.append(f"Captured stdout: {len(stdout_text)} chars")
-
-            # Find the last valid JSON line
-            last_json_line = None
-            for line in stdout_text.splitlines()[::-1]:
+            logs.append(f"Execution time: {time.time() - t0:.2f}s")
+            
+            # Parse JSON result
+            for line in reversed(stdout_text.splitlines()):
                 if JSON_LINE_RE.match(line.strip()):
-                    last_json_line = line.strip()
-                    break
-
-            if last_json_line:
-                try:
-                    parsed_json = json.loads(last_json_line)
-                    logger.info(f"[{self.AGENT_TYPE}] Successfully parsed JSON response")
-                except json.JSONDecodeError as e:
-                    logs.append(f"Failed to parse JSON line: {e}")
-
-            # Collect declared artifacts
-            artifacts_list = []
+                    try:
+                        parsed_json = json.loads(line.strip())
+                        logger.info(f"[{self.AGENT_TYPE}] Parsed JSON result")
+                        break
+                    except json.JSONDecodeError:
+                        pass
+            
+            # Collect artifacts
             if parsed_json and isinstance(parsed_json.get("artifacts"), list):
-                artifacts_list = parsed_json["artifacts"]
-
-            for rel_path in artifacts_list:
-                artifact_path = (workspace / rel_path).resolve()
-                try:
-                    # Ensure artifact is within workspace (security check)
-                    artifact_path.relative_to(workspace)
-                except ValueError:
-                    logs.append(f"Ignored artifact outside workspace: {rel_path}")
-                    continue
-
-                if artifact_path.is_file():
-                    artifacts.append({
-                        "name": str(Path(rel_path)),
-                        "content": artifact_path.read_bytes(),
-                    })
-                    logger.info(f"[{self.AGENT_TYPE}] Collected artifact: {rel_path}")
-                elif artifact_path.is_dir():
-                    logs.append(f"Directory artifact not packaged: {rel_path}")
-
-            # Determine final status
+                for rel_path in parsed_json["artifacts"]:
+                    artifact_path = (workspace_path / rel_path).resolve()
+                    try:
+                        artifact_path.relative_to(workspace_path)  # Security check
+                        if artifact_path.is_file():
+                            artifacts.append({
+                                "name": str(Path(rel_path)),
+                                "content": artifact_path.read_bytes(),
+                            })
+                            logger.info(f"[{self.AGENT_TYPE}] Collected artifact: {rel_path}")
+                    except (ValueError, FileNotFoundError) as e:
+                        logs.append(f"Artifact error: {rel_path}: {e}")
+            
+            # Determine status
             has_sentinel = SENTINEL in stdout_text
-            status = (parsed_json or {}).get("status") or ("ok" if has_sentinel else "error")
-            summary = (parsed_json or {}).get("summary") or ("completed" if status == "ok" else "incomplete")
-
+            status = (parsed_json or {}).get("status", "ok" if has_sentinel else "error")
+            summary = (parsed_json or {}).get("summary", "Task completed" if status == "ok" else "Task failed")
+            
             return {
                 "status": status,
                 "summary": summary,
                 "result": parsed_json or {},
                 "artifacts": artifacts,
                 "logs": logs,
-                "stdout": stdout_text,
+                "stdout": stdout_text[:10000] if len(stdout_text) > 10000 else stdout_text  # Limit for storage
             }
-
+            
         except Exception as e:
-            logger.error(f"[{self.AGENT_TYPE}] Exception in execution: {e}")
+            logger.error(f"[{self.AGENT_TYPE}] Execution failed: {e}")
             return {
                 "status": "error",
-                "summary": f"execution failed: {type(e).__name__}",
+                "summary": f"Execution failed: {type(e).__name__}",
                 "result": {},
                 "artifacts": [],
                 "logs": logs + [str(e)],
-                "stdout": "",
+                "stdout": ""
             }
-        finally:
-            # Always cleanup workspace
-            try:
-                shutil.rmtree(workspace, ignore_errors=True)
-                logger.debug(f"[{self.AGENT_TYPE}] Cleaned up workspace: {workspace}")
-            except Exception as e:
-                logger.warning(f"[{self.AGENT_TYPE}] Failed to cleanup workspace: {e}")
+    
+    async def _process_artifacts(
+        self, 
+        workspace: str, 
+        operation_id: str,
+        grantee_address: str,
+        result: Dict
+    ) -> List[Dict]:
+        """Process and store artifacts from agent execution."""
+        artifacts_metadata = []
+        
+        out_dir = os.path.join(workspace, "out")
+        if os.path.exists(out_dir):
+            for filename in os.listdir(out_dir):
+                filepath = os.path.join(out_dir, filename)
+                if os.path.isfile(filepath):
+                    with open(filepath, 'rb') as f:
+                        content = f.read()
+                    
+                    # Store in artifact storage
+                    await self.artifact_storage.store_artifact(
+                        operation_id, filename, content, grantee_address
+                    )
+                    
+                    # Add metadata
+                    artifacts_metadata.append({
+                        "name": filename,
+                        "size": len(content),
+                        "artifact_path": f"out/{filename}"
+                    })
+        
+        return artifacts_metadata
