+++ services/operations_refactored_incomplete.py
@@ -3,6 +3,7 @@ Refactored operations service following SOLID principles.
 Split into focused services with single responsibilities.
 """
 from typing import Dict, List, Optional, Any
+import asyncio
 import json
 import logging
